pyproject.toml
[tool.poetry]
name = "cca-zoo"
version = "0.0.0"
description = "Canonical Correlation Analysis Zoo: A collection of Regularized, Deep Learning based, Kernel, and Probabilistic methods in a scikit-learn style framework"
authors = ["jameschapman <james.chapman.19@ucl.ac.uk>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/jameschapman19/cca_zoo"
keywords = ["cca"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0.0"
numpy = "*"
scipy = "*"
scikit-learn = "^1.2.2"
pandas = "*"
tensorly = "*"
joblib = "*"
mvlearn = "*"
tqdm = "*"
torch = [
     {version = "^2.0.1", platform = "darwin"},
     {version = "^2.0.1", platform = "linux", source = "torch"},
     {version = "^2.0.1", platform = "win32", source = "torch"}, ]
lightning = "*"
# Optional dependencies
jax = {version = "*", optional = true}
numpyro = {version = "*", optional = true}
arviz = {version = "*", optional = true}
funsor = {version = "*", optional = true}

[[tool.poetry.source]]
 name = "torch"
 url = "https://download.pytorch.org/whl/cpu"
 priority = "explicit"

[tool.poetry.extras]
probabilistic = ["jax", "numpyro", "arviz", "funsor"]

[tool.poetry.dev-dependencies]
black = "*"
flake8 = "*"
codecov = "*"
pytest-cov = "*"
seaborn = "*"
opentsne = "*"
umap-learn = "*"
rdata = "*"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
